Work on making pose graph structure more generic; standardize on float/double

+++ src/base_lib/pose_reps.h
@@ -8,7 +8,7 @@
 #include <eigen3/Eigen/Dense>
 
 namespace pose {
-    typedef std::pair<Eigen::Vector2f, float> Pose2d;
+    typedef std::pair<Eigen::Vector2d, double> Pose2d;
     typedef std::pair<Eigen::Vector3d, Eigen::Quaterniond> Pose3d;
